Final steps Stage Manager app
