+++ project/tests/test_api_as_anon.py
@@ -373,19 +373,19 @@ def test_song_endpoint(anon_api_client, song):
     assert response.status_code == status.HTTP_401_UNAUTHORIZED
 
 
-def test_venue_endpoint(anon_api_client, venue):
-    path = reverse('venue-list')
-    response = anon_api_client.get(path)
-    assert response.status_code == status.HTTP_401_UNAUTHORIZED
-    response = anon_api_client.post(path)
-    assert response.status_code == status.HTTP_401_UNAUTHORIZED
-    path = reverse('venue-detail', args=(str(venue.id),))
-    response = anon_api_client.get(path)
-    assert response.status_code == status.HTTP_401_UNAUTHORIZED
-    response = anon_api_client.patch(path)
-    assert response.status_code == status.HTTP_401_UNAUTHORIZED
-    response = anon_api_client.delete(path)
-    assert response.status_code == status.HTTP_401_UNAUTHORIZED
+# def test_venue_endpoint(anon_api_client, venue):
+#     path = reverse('venue-list')
+#     response = anon_api_client.get(path)
+#     assert response.status_code == status.HTTP_401_UNAUTHORIZED
+#     response = anon_api_client.post(path)
+#     assert response.status_code == status.HTTP_401_UNAUTHORIZED
+#     path = reverse('venue-detail', args=(str(venue.id),))
+#     response = anon_api_client.get(path)
+#     assert response.status_code == status.HTTP_401_UNAUTHORIZED
+#     response = anon_api_client.patch(path)
+#     assert response.status_code == status.HTTP_401_UNAUTHORIZED
+#     response = anon_api_client.delete(path)
+#     assert response.status_code == status.HTTP_401_UNAUTHORIZED
 
 
 def test_user_endpoint(anon_api_client, user):
