+++ project/tests/test_api_as_user.py
@@ -213,13 +213,13 @@ def test_song_endpoint(user_api_client, song):
     assert response.status_code == status.HTTP_403_FORBIDDEN
 
 
-def test_venue_endpoint(user_api_client, venue):
-    path = reverse('venue-list')
-    response = user_api_client.get(path)
-    assert response.status_code == status.HTTP_200_OK
-    path = reverse('venue-detail', args=(str(venue.id),))
-    response = user_api_client.get(path)
-    assert response.status_code == status.HTTP_200_OK
+# def test_venue_endpoint(user_api_client, venue):
+#     path = reverse('venue-list')
+#     response = user_api_client.get(path)
+#     assert response.status_code == status.HTTP_200_OK
+#     path = reverse('venue-detail', args=(str(venue.id),))
+#     response = user_api_client.get(path)
+#     assert response.status_code == status.HTTP_200_OK
 
 
 def test_user_endpoint(user_api_client, user):
